requirements/prod.txt
Django==4.1.3
django_rest_swagger==2.2.0
djangorestframework==3.14.0
graphene==3.1.1
graphene_django==3.0.0